Split the web module into different files for receive mode and share mode logic
